+++ spec/unit/connection_spec.rb
@@ -1,48 +1,48 @@
 # encoding: utf-8
 
-describe ROM::Cassandra::Session do
+describe ROM::Cassandra::Connection do
 
-  let(:session) { described_class.new(uri) }
-  let(:uri)     { { hosts: ["127.0.0.1"], port: 9042 } }
-  let(:query)   { double to_s: "SELECT * FROM auth.users;" }
+  let(:conn)  { described_class.new(uri) }
+  let(:uri)   { { hosts: ["127.0.0.1"], port: 9042 } }
+  let(:query) { double to_s: "SELECT * FROM auth.users;" }
 
   describe ".new" do
-    subject { session }
+    subject { conn }
     let(:uri) { { hosts: ["127.0.0.1"], port: 1000 } }
 
     it "fails with wrong uri" do
-      expect { session }.to raise_error StandardError
+      expect { subject }.to raise_error StandardError
     end
   end
 
   describe "#uri" do
-    subject { session.uri }
+    subject { conn.uri }
 
     context "from hash" do
       it { is_expected.to eql uri }
     end
 
     context "from nil" do
-      let(:session) { described_class.new }
+      let(:conn) { described_class.new }
 
       it { is_expected.to eql uri }
     end
 
     context "from string" do
-      let(:session) { described_class.new "127.0.0.1:9042" }
+      let(:conn) { described_class.new "127.0.0.1:9042" }
 
       it { is_expected.to eql uri }
     end
 
     context "from string and hash" do
-      let(:session) { described_class.new "127.0.0.1", port: 9042 }
+      let(:conn) { described_class.new "127.0.0.1", port: 9042 }
 
       it { is_expected.to eql uri }
     end
   end # describe #uri
 
   describe "#call" do
-    subject { session.call(query) }
+    subject { conn.call(query) }
 
     it "returns the result of running query" do
       expect(subject).to eql [
@@ -52,4 +52,4 @@ describe ROM::Cassandra::Session do
     end
   end # describe #call
 
-end # describe ROM::Cassandra::Session
+end # describe ROM::Cassandra::Connection